+++ app/src/main/AndroidManifest.xml
@@ -70,11 +70,6 @@
         </activity>
 
         <activity
-            android:name=".ui.AboutActivity"
-            android:label="@string/navigation_drawer_about"
-            android:parentActivityName=".ui.MainActivity"/>
-
-        <activity
             android:name=".ui.AppSettingsActivity"
             android:label="@string/app_settings"
             android:parentActivityName=".ui.MainActivity"
