+++ app/src/main/kotlin/at/bitfire/davdroid/ui/about/AppLicenseInfoProvider.kt
@@ -1,0 +1,8 @@
+package at.bitfire.davdroid.ui.about
+
+import androidx.compose.runtime.Composable
+
+interface AppLicenseInfoProvider {
+    @Composable
+    fun LicenseInfo()
+}
