+++ app/src/ose/kotlin/at/bitfire/davdroid/ui/OpenSourceLicenseInfoProvider.kt
@@ -14,17 +14,18 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.core.text.HtmlCompat
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
-import androidx.lifecycle.viewmodel.compose.viewModel
 import at.bitfire.davdroid.ui.UiUtils.toAnnotatedString
+import at.bitfire.davdroid.ui.about.AppLicenseInfoProvider
 import com.google.common.io.CharStreams
 import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
+import javax.inject.Inject
 
-class OpenSourceLicenseInfoProvider @Inject constructor(): AboutActivity.AppLicenseInfoProvider {
+class OpenSourceLicenseInfoProvider @Inject constructor(): AppLicenseInfoProvider {
 
     @Composable
     override fun LicenseInfo() {
@@ -33,7 +34,7 @@ class OpenSourceLicenseInfoProvider @Inject constructor(): AboutActivity.AppLice
 
     @Composable
     fun LicenseInfoGpl(
-        model: Model = viewModel()
+        model: Model = hiltViewModel()
     ) {
         model.gpl?.let { OpenSourceLicenseInfo(it.toAnnotatedString()) }
     }
